Moved "useInfiniteGallery" out
